HW 7. extracted some functions to separate files
