+++ HW_5/OOP-functions/oop/Functions/NewTaskOnEnter.ts
@@ -1,0 +1,4 @@
+export const NewTaskOnEnter = () => {
+        const newItemBox: HTMLButtonElement = document.getElementById('NewItemButtonApply') as HTMLButtonElement;
+        newItemBox.disabled == false ? newItemBox.onclick.call(this) : '';
+    }
