+++ HW_5/OOP-functions/oop/Functions/SearchByPattern.ts
@@ -1,0 +1,10 @@
+import ItemInterface from '../Interfaces/ItemInterface';
+
+export function SearchByPattern(taskList: ItemInterface[], searchPattern: string) {
+    for (let i in taskList) {
+        taskList[i].htmlElement.className = ("tasks__row");
+        if (!taskList[i].name.match(searchPattern)){
+            taskList[i].htmlElement.className = ("tasks__row tasks__row--disabled");
+        }  
+    }
+}
