+++ HW_5/OOP-functions/oop/Functions/index.ts
@@ -1,0 +1,10 @@
+export * from './UpdateDataOnServer.ts';
+export * from './DeleteDataFromServer.ts';
+export * from './GetDataFromServer.ts';
+export * from './PostNewDataIntoServer.ts';
+export * from './SearchByPattern.ts';
+export * from './MorningGreatings.ts';
+export * from './SearchOnLoad.ts';
+export * from './AproveNewItem.ts';
+export * from './NewTaskOnEnter.ts';
+export * from './CreateListForMorningGreeting.ts';
