+++ HW_5/OOP-functions/oop/app.ts
@@ -6,9 +6,11 @@ import TaskRowUncompleted from './Components/TaskRowUncompleted.ts';
 
 
 import Properities from './Interfaces/Properities'; 
-import StateIterface from './Interfaces/StateInterface'; 
-import ItemInterface from './Interfaces/ItemInterface'
-import HTMLCommonElement from './Interfaces/HTMLCommonElement'
+import StateInterface from './Interfaces/StateInterface'; 
+import ItemInterface from './Interfaces/ItemInterface';
+import HTMLCommonElement from './Interfaces/HTMLCommonElement';
+
+import * as Functions from './Functions';
 
 import './app.css';
 
@@ -18,6 +20,8 @@ class App extends Component {
 
     constructor() {
         super();
+        let taskList: ItemInterface[];
+
         this.CheckServerAvailability(config.localurl)
             .then((isAvailable: boolean) => {
                 if (isAvailable) {
@@ -30,7 +34,12 @@ class App extends Component {
                 localStorage.setItem("server_url", config.githuburl);
             })
             .finally(() => {
-                this.GetDataFromServer();
+                Functions.GetDataFromServer()
+                    .then((data: ItemInterface[]) => {
+                        this.setState({
+                            taskList: data
+                        });
+                    });
             })
 
         this.setState({
@@ -54,9 +63,9 @@ class App extends Component {
 
         this.WeatherCall();
 
-        this.MorningGreatings(this.state.currentDate);
+        Functions.MorningGreatings(this.state.currentDate);
 
-        this.SearchOnLoad();
+        Functions.SearchOnLoad(this.state.taskList);
         
         return super.render({
             id: 'oop_container',
@@ -143,32 +152,6 @@ class App extends Component {
         } 
     }
 
-    MorningGreatings = (currentDate: string) => {
-        if (localStorage.getItem("currentDate")) {
-            if (localStorage.getItem("currentDate") != currentDate) {
-                localStorage.setItem("currentDate", currentDate);
-                localStorage.setItem("modalWindowState", '1');
-            }
-        } else {
-            localStorage.setItem("currentDate", currentDate);
-            localStorage.setItem("modalWindowState", '1');
-        }
-    }
-
-    SearchOnLoad = () => {
-        let searchPattern: string = '';
-        if (localStorage.getItem("searchPattern")) {
-            searchPattern = localStorage.getItem("searchPattern");
-        };
-
-        for (let i in this.state.taskList) {
-            this.state.taskList[i].htmlElement.className = ("tasks__row");
-            if (!this.state.taskList[i].name.match(searchPattern)){
-                this.state.taskList[i].htmlElement.className = ("tasks__row tasks__row--disabled");
-            }  
-        }
-    }
-
     ApplyItem = () => {
         const taskList: ItemInterface[] = this.state.taskList;
         const newItemInput: HTMLInputElement = document.getElementById("NewItemInput") as HTMLInputElement;
@@ -191,7 +174,7 @@ class App extends Component {
             tag: localStorage.getItem('currentTag')
         };
 
-        this.PostNewDataIntoServer(newTask);
+        Functions.PostNewDataIntoServer(newTask);
 
         localStorage.setItem("modalWindowState", '0');
         localStorage.setItem("currentTag", 'other');
@@ -208,7 +191,7 @@ class App extends Component {
         const pickedElementIndex: number = taskList.indexOf(taskList.find(task => task.htmlElement == parent));
         const removedElement: ItemInterface[] = taskList.splice(pickedElementIndex, 1);
 
-        this.DeleteDataFromServer(removedElement[0]);
+        Functions.DeleteDataFromServer(removedElement[0]);
 
         this.setState(
             {taskList: [...taskList]},
@@ -231,8 +214,6 @@ class App extends Component {
         const newItemInput: HTMLInputElement = document.getElementById('NewItemInput') as HTMLInputElement;
         newItemInput.value = '';
         newItemInput.focus();
-
-        
     }
 
     CreateListOfElements = (itemList: ItemInterface[]) => {
@@ -271,82 +252,9 @@ class App extends Component {
         let pickedElement = this.state.taskList.find(task => task.htmlElement === parent);
 
         pickedElement.isCompleted = !pickedElement.isCompleted;
-        this.UpdateDataOnServer(pickedElement);
+        Functions.UpdateDataOnServer(pickedElement);
 
         this.update();
-    }
-
-    GetDataFromServer = async () => {
-        let taskList: ItemInterface[] = [];
-        const serverUrl: string = localStorage.getItem("server_url");
-
-        if (serverUrl) {
-            await fetch(serverUrl, { method: "GET" })
-                .then((data) => data.json())
-                .then((data) => {
-                    if (Array.isArray(data)) {
-                        for (let item of data) {
-                            taskList.push({
-                                name: item.title, 
-                                id: item.id, 
-                                isCompleted: item.isCompleted,
-                                plannedDate: item.plannedDate,
-                                tag: item.tag
-                            });
-                        }
-                    }
-                })
-                .catch((error) => {window.alert(error)})
-
-            this.setState({
-                taskList: taskList,
-            });
-        }
-    }
-
-    PostNewDataIntoServer = async (data: ItemInterface) => {
-        const serverUrl: string = localStorage.getItem("server_url");
- 
-        await fetch(serverUrl, { 
-            method: "POST", 
-            headers: {
-                'Content-type': 'application/json; charset=UTF-8',
-            },
-            body: JSON.stringify({
-                id: data.id, 
-                title:data.name, 
-                isCompleted: data.isCompleted,
-                plannedDate: data.plannedDate,
-                tag: data.tag
-            })
-            })
-            .catch((error) => window.alert(error));
-    }
-
-    DeleteDataFromServer = async (data: ItemInterface) => {
-        const serverUrl: string = localStorage.getItem("server_url");
-
-        await fetch(serverUrl + '\/' + data.id, {method: "DELETE"})
-            .catch((error) => window.alert(error));
-    }
-
-    UpdateDataOnServer = async (data: ItemInterface) => {
-        const serverUrl: string = localStorage.getItem("server_url");
-
-        await fetch(serverUrl + '\/' + data.id, { 
-            method: "PUT", 
-            headers: {
-                'Content-type': 'application/json; charset=UTF-8',
-            },
-            body: JSON.stringify({
-                id: data.id, 
-                title:data.name, 
-                isCompleted: data.isCompleted,
-                plannedDate: data.plannedDate,
-                tag: data.tag
-            })
-            })
-            .catch((error) => window.alert(error));
     }
 
     CheckServerAvailability = async (url: string): Promise<boolean> => {
